+++ R/read_params.R
@@ -33,7 +33,7 @@ read_params <- function(stage_path = NULL, return_list = FALSE) {
     if (is.null(config_env$stage_dir)) {
       stop("stage_path argument missing.")
     } else {
-      cat(paste("reloading from already set stage_path:", config_env$stage_dir))
+      message(glue::glue("Reloading from already set stage_path: {config_env$stage_dir}"))
       stage_path <- config_env$stage_dir
     }
   }